feat: refactor content-runtime to include new approach from other content scripts
