--- pages/content-runtime/src/index.ts
@@ -1,4 +1,0 @@
-import { mount } from '@src/index.tsx';
-
-mount();
-console.log('runtime script loaded');
