+++ pages/content-runtime/src/matches/example/index.ts
@@ -1,0 +1,4 @@
+import { mount } from '@src/matches/example/Root';
+
+mount();
+console.log('runtime script loaded');
